1. 리팩토링: Rest API로 변경
2. 리팩토링: 게시판 엔드포인트를 /board로 수정 및 다수의 JSP 파일명을 게시판 관련 파일명(boards, board, write, update, reply)으로 일괄 수정

3. 리팩토링: 전체 테스트 코드 수정

4. 기능 변경: 소셜 로그인(네이버, 카카오)을 REST API 방식에서 자바스크립트 방식으로 변경

+++ jam/src/main/java/com/jam/client/comReply/dao/ComReplyDAO.java
@@ -2,14 +2,13 @@ package com.jam.client.comReply.dao;
 
 import java.util.List;
 
-import org.apache.ibatis.annotations.Param;
 
 import com.jam.client.comReply.vo.ComReplyVO;
 
 public interface ComReplyDAO {
 
 	// 커뮤니티 댓글 리스트
-	public List<ComReplyVO> replyList(Integer com_no);
+	public List<ComReplyVO> replyList(Long com_no);
 
 	//커뮤니티 댓글 입력
 	public int replyInsert(ComReplyVO crvo);
@@ -18,10 +17,10 @@ public interface ComReplyDAO {
 	public int replyUpdate(ComReplyVO crvo);
 
 	// 커뮤니티 댓글 삭제
-	public int replyDelete(Integer comReply_no);
+	public int replyDelete(Long comReply_no);
 
 	// 커뮤니티 댓글의 글 번호 조회
-	public ComReplyVO replyRead(Integer comReply_no);
+	public ComReplyVO replyRead(Long comReply_no);
 
 	
 }
