+++ jam/src/main/java/com/jam/client/comReply/vo/ComReplyVO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 
 @Data
 public class ComReplyVO {
-	private int comReply_no; 			// 커뮤니티 댓글 번호
-	private int com_no; 				// 커뮤니티 글 번호
+	private Long comReply_no; 			// 커뮤니티 댓글 번호
+	private Long com_no; 				// 커뮤니티 글 번호
 	private String comReply_content; 	// 커뮤니티 댓글 내용
 	private String comReply_date; 		// 커뮤니티 댓글 작성일
 	private String user_id; 			// 커뮤니티 댓글 작성자 id
