+++ jam/src/main/java/com/jam/client/fleaReply/dao/FleaReplyDAO.java
@@ -7,7 +7,7 @@ import com.jam.client.fleaReply.vo.FleaReplyVO;
 public interface FleaReplyDAO {
 
 	// 중고악기 댓글 리스트
-	public List<FleaReplyVO> replyList(Integer flea_no);
+	public List<FleaReplyVO> replyList(Long flea_no);
 
 	// 중고악기 댓글 입력
 	public int replyInsert(FleaReplyVO frvo);
@@ -16,9 +16,9 @@ public interface FleaReplyDAO {
 	public int replyUpdate(FleaReplyVO frvo);
 	
 	// 중고악기 댓글 삭제
-	public int replyDelete(Integer fleaReply_no);
+	public int replyDelete(Long fleaReply_no);
 	
 	// 중고악기 댓글의 글 번호 조회
-	public FleaReplyVO replyRead(Integer fleaReply_no);
+	public FleaReplyVO replyRead(Long fleaReply_no);
 
 }
