+++ jam/src/main/java/com/jam/client/jobReply/dao/JobReplyDAO.java
@@ -7,7 +7,7 @@ import com.jam.client.jobReply.vo.JobReplyVO;
 public interface JobReplyDAO {
 
 	// 구인구직 댓글 리스트
-	public List<JobReplyVO> replyList(Integer job_no);
+	public List<JobReplyVO> replyList(Long job_no);
 
 	// 구인구직 댓글 입력
 	public int replyInsert(JobReplyVO jrvo);
@@ -16,9 +16,9 @@ public interface JobReplyDAO {
 	public int replyUpdate(JobReplyVO jrvo);
 
 	// 구인구직 댓글 삭제
-	public int replyDelete(Integer jobReply_no);
+	public int replyDelete(Long jobReply_no);
 
 	// 구인구직 댓글의 글 번호 조회
-	public JobReplyVO replyRead(Integer jobReply_no);
+	public JobReplyVO replyRead(Long jobReply_no);
 
 }
