+++ jam/src/main/java/com/jam/client/roomRentalReply/service/RoomReplyService.java
@@ -7,7 +7,7 @@ import com.jam.client.roomRentalReply.vo.RoomReplyVO;
 public interface RoomReplyService {
 	
 	// 댓글 리스트
-	List<RoomReplyVO> roomReplyList(Integer roomRental_no);
+	List<RoomReplyVO> roomReplyList(Long roomRental_no);
 
 	// 댓글 입력
 	int replyInsert(RoomReplyVO rrvo);
@@ -16,6 +16,6 @@ public interface RoomReplyService {
 	int replyUpdate(RoomReplyVO rrvo);
 
 	// 댓글 삭제
-	int replyDelete(Integer roomReply_no);
+	int replyDelete(Long roomReply_no);
 
 }
